Incorporated suggested changes
